Fix bugs in path destruction and repair

+++ md5/md5_dispatch.c
@@ -22,5 +22,6 @@ int		index_of(char *s, int x)
 	init_md5(&set);
 	update_md5(&set, (t_ptr)s, len);
 	end_md5(digest, &set);
-	return(((digest[0] * digest[15]) ^ (digest[0] + digest[15])) % x);
+	return((((unsigned int)digest[0] * digest[15]) ^
+			((unsigned int)digest[0] + digest[15])) % x);
 }
